IOCP 后端与 Windows 平台层；MSVC 作业加入 CI
后端（src/detail/iocp，Corosio win_scheduler 的形状）：
- iocp_backend：一个完成端口；run() 阻塞在 GetQueuedCompletionStatus 上（超时取定时器堆顶），之后
  零超时排空；interrupt() 用 PostQueuedCompletionStatus 空包；定时器复用新抽出的 timer_heap /
  heap_timer（与 reactor 共用）；信号自管道读端做常驻重叠 WSARecv（带不满一个 int 的残留处理）
- iocp_socket：WSARecv / WSASend / WSARecvFrom / WSASendTo / ConnectEx（WSAIoctl 取指针、未绑定先绑
  通配地址、完成后 SO_UPDATE_CONNECT_CONTEXT）/ AcceptEx（预建对端、SO_UPDATE_ACCEPT_CONTEXT）；不做
  投机，发起即提交，同步失败直接以错误恢复；取消 CancelIoEx，closesocket 让在飞操作以 995 完成
- iocp_file：FILE_FLAG_OVERLAPPED 句柄上 ReadFile / WriteFile 带偏移；ERROR_HANDLE_EOF → eof；append 用
  "只有 FILE_APPEND_DATA" 的访问权限表达
- iocp_error：Win32 / WSA 两段错误码统一映射到 std::errc（MSVC 与 libstdc++ 的 system_category 各映射
  一段，这样两边都能与 std::errc 比较）

可移植层：
- net/detail/socket_types.hpp：native_socket_type / native_file_type / invalid_socket / shutdown_* 与
  平台套接字头；config.hpp 的 NET_PLATFORM_WINDOWS / NET_PLATFORM_LINUX；backend.hpp 的 iocp 标签
  （Windows 默认后端）
- 接缝 socket_impl / file_impl / register_signal_reader 用平台句柄类型；detail/file_ops.hpp 分出文件
  同步操作（posix/file_ops.cpp、iocp/win_file_ops.cpp 两份实现）
- socket_base / tcp / udp / resolver / ip / signal_set 去 POSIX 化：closesocket / ioctlsocket / WSAStartup
  （ensure_networking_initialized）、Windows 信号用 signal() + 回环套接字对
- local.hpp 暂不支持 Windows（#error），net.hpp 按平台包含
- 定时器堆抽成 detail/timer_heap + heap_timer，reactor_timer 删除

构建 / 测试：
- CMake 按平台选源码；Windows 链 ws2_32 / mswsock，PUBLIC 的 /permissive- /Zc:preprocessor /utf-8 /bigobj；
  测试只编 iocp 一个变体，local_tests 不编
- tests/platform.hpp：裸套接字 / 信号号（Windows 用 SIGINT / SIGTERM）/ 临时文件的可移植封装；
  backlog-0 的 connect 回归只在 POSIX 跑；只读文件写入接受 permission_denied
- 本地用 llvm-mingw（clang 21 + mingw-w64 UCRT 头）交叉编译库、15 个测试与 4 个示例全部通过 -Werror
- CI 新增 msvc 作业（windows-latest × Debug / Release，TLS 关）；cl 的诊断经 problem matcher、ctest 失败
  经 annotate_ctest.py 变成注解，公开仓库匿名可读

修复：any_* 整序列操作 std::move 与 buffer_size 同列实参的求值顺序问题（clang 作业失败的原因）。

+++ include/net/resolver.hpp
@@ -7,8 +7,7 @@
 #include <system_error>
 #include <vector>
 
-#include <netdb.h>
-#include <sys/socket.h>
+#include "net/detail/socket_types.hpp"
 
 #include "net/coroutine.hpp"
 #include "net/io_env.hpp"
@@ -126,8 +125,8 @@ template <class Protocol> struct basic_resolver_awaitable {
         entries.reserve(raw.size());
         for (auto& entry : raw) {
             auto endpoint = endpoint_type{};
-            if (entry.length <= endpoint.capacity()) {
-                std::memcpy(endpoint.data(), &entry.storage, entry.length);
+            if (static_cast<std::size_t>(entry.length) <= endpoint.capacity()) {
+                std::memcpy(endpoint.data(), &entry.storage, static_cast<std::size_t>(entry.length));
             }
             entries.push_back(basic_resolver_entry<Protocol>{endpoint, std::move(entry.host_name),
                                                              std::move(entry.service_name)});
